*** empty log message ***

+++ src/services/gov/epa/emissions/framework/services/cost/analysis/leastcost/StrategyTask.java
@@ -16,6 +16,7 @@ import gov.epa.emissions.framework.services.cost.ControlStrategyInputDataset;
 import gov.epa.emissions.framework.services.cost.analysis.common.AbstractStrategyTask;
 import gov.epa.emissions.framework.services.cost.controlStrategy.ControlStrategyResult;
 import gov.epa.emissions.framework.services.cost.controlStrategy.DatasetCreator;
+import gov.epa.emissions.framework.services.cost.controlStrategy.StrategyResultType;
 import gov.epa.emissions.framework.services.data.DatasetTypesDAO;
 import gov.epa.emissions.framework.services.data.EmfDataset;
 import gov.epa.emissions.framework.services.persistence.HibernateSessionFactory;
@@ -57,8 +58,18 @@ public class StrategyTask extends AbstractStrategyTask {
 //        super.run(loader);
         
         //get rid of strategy results
-        deleteStrategyResults();
-
+//        deleteStrategyResults();
+        if (controlStrategy.getDeleteResults()) {
+//            List<ControlStrategyResult> resultList = new ArrayList<ControlStrategyResult>(); 
+//            ControlStrategyResult[] results = getControlStrategyResults();
+//            for (ControlStrategyResult result : results) {
+//                if (result.getStrategyResultType().getName().equals(StrategyResultType.leastCostControlMeasureWorksheetResult))
+//                resultList.add(result);
+//            }
+//          deleteStrategyResults(resultList.toArray(new ControlStrategyResult[0]));
+          deleteStrategyResults();
+        }
+        
         //run any pre processes
         try {
             beforeRun();
@@ -125,16 +136,38 @@ public class StrategyTask extends AbstractStrategyTask {
     }
 
     public void beforeRun() throws EmfException {
-        //create the worksheet (strat result)
-        leastCostCMWorksheetResult = loader.loadLeastCostCMWorksheetResult();
-        
+        ControlStrategyResult[] results = getControlStrategyResults();
+        //create the worksheet (strat result), if needed, maybe they don't want to recreate these...
+        if (controlStrategy.getDeleteResults() || results.length == 0) {
+            leastCostCMWorksheetResult = loader.loadLeastCostCMWorksheetResult();
+        } else {
+            for (ControlStrategyResult result : results ) {
+                if (result.getStrategyResultType().getName().equals(StrategyResultType.leastCostControlMeasureWorksheetResult)) {
+                    leastCostCMWorksheetResult = result;
+                    break;
+                }
+            }
+        }
         //if there is more than one input inventory, then merge these into one dataset, 
         //then we use that as the input to the strategy run
-        if (controlStrategyInputDatasetCount >= 1) {
-            //check to see if exists already, if so, then truncate its data and start over...
+        if (controlStrategy.getControlStrategyInputDatasets().length > 1) {
+//        if (controlStrategyInputDatasetCount >= 1) {
             ControlStrategyInputDataset[] inputDatasets = controlStrategy.getControlStrategyInputDatasets();
+
+            //TODO: look for any errors or warnings in the input inventories
+            //i.e., missing sector or differing temporal period
+//            String sector = getDatasetSector(inputDatasets[0].getInputDataset());
+//            for (ControlStrategyInputDataset inputDataset : inputDatasets) 
+//                if (!inputDataset.getInputDataset().getDatasetType().getName().equals(DatasetType.orlMergedInventory)) {
+//                    inputDataset.getInputDataset().getTemporalResolution()
+//                    hasMergedDataset = true;
+//                    mergedDataset = inputDataset.getInputDataset();
+//                }
+            
+            //check to see if exists already, if so, then truncate its data and start over...
             boolean hasMergedDataset = false;
             EmfDataset mergedDataset = null;
+            //see if it already has a merged dataset
             for (ControlStrategyInputDataset inputDataset : inputDatasets) 
                 if (inputDataset.getInputDataset().getDatasetType().getName().equals(DatasetType.orlMergedInventory)) {
                     hasMergedDataset = true;
